[57-start-end-functionality] defaultGameSettings перенесены в gameSlice
